[CHORE] Break down generic components into multiple components
